+++ internal/provider/model_data_source_meraki_switch_ports.go
@@ -58,6 +58,7 @@ type DataSourceSwitchPortsItems struct {
 	StickyMacAllowListLimit                types.Int64  `tfsdk:"sticky_mac_allow_list_limit"`
 	StormControlEnabled                    types.Bool   `tfsdk:"storm_control_enabled"`
 	StpGuard                               types.String `tfsdk:"stp_guard"`
+	StpPortFastTrunk                       types.Bool   `tfsdk:"stp_port_fast_trunk"`
 	Type                                   types.String `tfsdk:"type"`
 	Udld                                   types.String `tfsdk:"udld"`
 	Vlan                                   types.Int64  `tfsdk:"vlan"`
@@ -190,6 +191,11 @@ func (data *DataSourceSwitchPorts) fromBody(ctx context.Context, res meraki.Res)
 			data.StpGuard = types.StringValue(value.String())
 		} else {
 			data.StpGuard = types.StringNull()
+		}
+		if value := res.Get("stpPortFastTrunk"); value.Exists() && value.Value() != nil {
+			data.StpPortFastTrunk = types.BoolValue(value.Bool())
+		} else {
+			data.StpPortFastTrunk = types.BoolNull()
 		}
 		if value := res.Get("type"); value.Exists() && value.Value() != nil {
 			data.Type = types.StringValue(value.String())