+++ docs/guides/changelog.md
@@ -14,6 +14,9 @@ description: |-
 - Add `sso_login_url` attribute to `meraki_organization_saml_idp` resources and data sources
 - Add `uplink_selection_candidates` and `uplink_selection_failback_enabled` attributes to `meraki_switch_settings` resource and data source
 - Add `meraki_wireless_ssid_open_roaming` resource and data source
+- Add `sp_initiated_idp_id` and `sp_initiated_subdomain` attributes to `meraki_organization_saml` resource and data source
+- Add `stp_port_fast_trunk` attribute to `meraki_switch_port` resources and data sources
+- Add `access_control_*` attributes to `meraki_wireless_ssids` data source
 
 ## 1.9.0